// src/models/ProductModel.ts
export interface ProductModel {
  id: string;
  idProvider: string;
  code: string;
  description: string;
  unity: string;
  price: number;
  group: string;
  subgroup: string;
  active: string;
  fractioned: string;
  fractions: number;
  edge: string;
  additional: string;
  portionSize: string;
  web_img_64: string;
  web_borda_subgrupo: string;
  monday: string;
  monday_start: string;
  monday_stop: string;
  tuesday: string;
  tuesday_start: string;
  tuesday_stop: string;
  wednesday: string;
  wednesday_start: string;
  wednesday_stop: string;
  thursday: string;
  thursday_start: string;
  thursday_stop: string;
  friday: string;
  friday_start: string;
  friday_stop: string;
  saturday: string;
  saturday_start: string;
  saturday_stop: string;
  sunday: string;
  sunday_start: string;
  sunday_stop: string;
}